requirements/base.txt
# Requirements for Lecture Capture Preferences itself
django>=2.0
psycopg2-binary
# explicitly specify django-automationcommon's git repo since changes in
# automationcommon tend to be "ad hoc" and may need testing here without a
# corresponding pypi release. Recall that git branched may be explicitly given
# in the VCS URL.
git+https://github.com/uisautomation/django-automationcommon@master#egg=django-automationcommon
django-ucamwebauth>=1.4.5
django-ucamlookup

# We need at least version 4 of whitenoise to make use of the index_file
# configuration option.
whitenoise>=4.0

# API-related Django packages
djangorestframework
django-cors-headers
drf-yasg
django-crispy-forms
django-filter

# Required for loading fixtures from YAML files
PyYAML

# For an improved manage.py shell experience
ipython

# So that tests may be run within the container
tox

# Serving
gunicorn
